Revert "updated with offline, hardware-wallet and multiwitness"
This reverts commit af9d0ffff49d0e8adbde97132e2ef64f30f49b5e.

+++ cardano/testnet-launchpad/01_sendAssets.sh
@@ -1,76 +1,37 @@
 #!/bin/bash
 
-# Script is brought to you by ATADA_Stakepool, Telegram @atada_stakepool in
-# cooperation with @TheRealAdamDean (BUFFY & SPIKE)
+# Script is brought to you by ATADA_Stakepool, Telegram @atada_stakepool
 
 #load variables from common.sh
 #       socket          Path to the node.socket (also exports socket to CARDANO_NODE_SOCKET_PATH)
 #       genesisfile     Path to the genesis.json
 #       magicparam      TestnetMagic parameter
 #       cardanocli      Path to the cardano-cli executable
+#       cardanonode     Path to the cardano-node executable
 . "$(dirname "$0")"/00_common.sh
 
 case $# in
-  2 ) fromAddr="$(dirname $1)/$(basename $1 .addr)"; fromAddr=${fromAddr//.\//}
-      metafile="$(dirname $2)/$(basename $2 .json).json"; metafile=${metafile//.\//};;
+  4|5 ) fromAddr="$1";
+      toAddr="$2";
+      assetToSend="$3";
+      amountToSend="$4";;
   * ) cat >&2 <<EOF
-Usage:  $(basename $0) <From AddressName> <VoteFileName>.json
+Usage:  $(basename $0) <From AddressName> <To AddressName or HASH> <PolicyID.Name of the asset OR <PATHtoNAME>.asset of the asset file> <Amount of assets to send OR keyword ALL> [Optional Amount of lovelaces to attach, normally zero]
 EOF
   exit 1;; esac
 
-#Check about the needed files
+if [[ $# -eq 5 ]]; then lovelacesToSend="$5"; else lovelacesToSend=0; fi
+
 if [ ! -f "${fromAddr}.addr" ]; then echo -e "\n\e[35mERROR - \"${fromAddr}.addr\" does not exist! Please create it first with script 03a or 02.\e[0m"; exit 1; fi
-if ! [[ -f "${fromAddr}.skey" || -f "${fromAddr}.hwsfile" ]]; then echo -e "\n\e[35mERROR - \"${fromAddr}.skey/hwsfile\" does not exist! Please create it first with script 03a or 02.\e[0m"; exit 1; fi
+if [ ! -f "${fromAddr}.skey" ]; then echo -e "\n\e[35mERROR - \"${fromAddr}.skey\" does not exist! Please create it first with script 03a or 02.\e[0m"; exit 1; fi
 
-#This is a simplified Version of the sendLovelaces.sh script so, it will always be a SendALLLovelaces transaction + Metafile
-toAddr=${fromAddr}
-lovelacesToSend="ALL"
+#Check if toAddr file doesn not exists, make a dummy one in the temp directory and fill in the given parameter as the hash address
+if [ ! -f "$2.addr" ]; then echo "$2" > ${tempDir}/tempTo.addr; toAddr="${tempDir}/tempTo"; fi
 
-#Throw an error if the voting.json file does not exist
-if [ ! -f "${metafile}" ]; then
-  echo -e "The specified VoteFileName.json (${metafile} file does not exist. Please try again."
-  exit 1
-fi
+#Check if the assetToSend is a file xxx.asset then read out the data from the file instead
+if [ -f "${3}.asset" ]; then assetToSend="$(jq -r .policyID < ${3}.asset).$(jq -r .name < ${3}.asset)"; fi
 
-function readMetaParam() {
-  required="${3:-0}"
-  key=$(jq 'keys[0]' $2)
-  param=$(jq -r ".$key .$1" $2 2> /dev/null)
-  if [[ $? -ne 0 ]]; then echo "ERROR - ${2} is not a valid JSON file" >&2; exit 1;
-  elif [[ "${param}" == null && required -eq 1 ]]; then echo "ERROR - Parameter \"$1\" in ${2} does not exist" >&2; exit 1;
-  elif [[ "${param}" == "" && !required -eq 1 ]]; then echo "ERROR - Parameter \"$1\" in ${2} is empty" >&2; exit 1;
-  fi
-  echo "${param}"
-}
-
-objectType=$(readMetaParam "ObjectType" "${metafile}" 1); if [[ ! $? == 0 ]]; then exit 1; fi
-objectVersion=$(readMetaParam "ObjectVersion" "${metafile}"); if [[ ! $? == 0 ]]; then exit 1; fi
-
-if [[ $objectType == 'VoteBallot' ]]; then
-  # Check VoteBallot required fields
-  networkId=$(readMetaParam "NetworkId" "${metafile}" 1); if [[ ! $? == 0 ]]; then exit 1; fi
-  proposalId=$(readMetaParam "ProposalId" "${metafile}" 1); if [[ ! $? == 0 ]]; then exit 1; fi
-  voterId=$(readMetaParam "VoterId" "${metafile}" 1); if [[ ! $? == 0 ]]; then exit 1; fi
-  yesnovote=$(readMetaParam "Vote" "${metafile}"); if [[ ! $? == 0 ]]; then exit 1; fi
-  choicevote=$(readMetaParam "Choices" "${metafile}"); if [[ ! $? == 0 ]]; then exit 1; fi
-  if [[ $yesnovote == null && $choicevote == null ]]; then
-    echo "ERROR - No voting preferences found in ballot.";
-    exit 1;
-  fi
-else
-    echo "ERROR - JSON is not of type VoteBallot.";
-    exit 1;
-fi
-
-sendFromAddr=$(cat ${fromAddr}.addr)
-sendToAddr=$(cat ${toAddr}.addr)
-check_address "${sendFromAddr}"
-check_address "${sendToAddr}"
-
-rxcnt="1"
-
-echo
-echo -e "\e[0mUsing lovelaces from Address\e[32m ${fromAddr}.addr\e[0m to send the metafile\e[32m ${metafile}\e[0m:"
+echo -e "\e[0mSending assets from Address\e[32m ${fromAddr}.addr\e[0m to Address\e[32m ${toAddr}.addr\e[0m:"
 echo
 
 #get live values
@@ -80,11 +41,17 @@ currentEPOCH=$(get_currentEpoch)
 
 echo -e "\e[0mCurrent Slot-Height:\e[32m ${currentTip} \e[0m(setting TTL[invalid_hereafter] to ${ttl})"
 echo
-echo -e "\e[0mSource/Destination Address ${fromAddr}.addr:\e[32m ${sendFromAddr} \e[90m"
-echo -e "\e[0mAttached Metafile:\e[32m ${metafile} \e[90m"
+
+sendFromAddr=$(cat ${fromAddr}.addr)
+sendToAddr=$(cat ${toAddr}.addr)
+
+check_address "${sendFromAddr}"
+check_address "${sendToAddr}"
+
+echo -e "\e[0mSource Address ${fromAddr}.addr:\e[32m ${sendFromAddr} \e[90m"
+echo -e "\e[0mDestination Address ${toAddr}.addr:\e[32m ${sendToAddr} \e[90m"
 echo
 
-#-------------------------------------
 #
 # Checking UTXO Data of the source address and gathering data about total lovelaces and total assets
 #
@@ -100,22 +67,19 @@ echo
 	txcnt=$(jq length <<< ${utxoJSON}) #Get number of UTXO entries (Hash#Idx), this is also the number of --tx-in for the transaction
 	if [[ ${txcnt} == 0 ]]; then echo -e "\e[35mNo funds on the Source Address!\e[0m\n"; exit; else echo -e "\e[32m${txcnt} UTXOs\e[0m found on the Source Address!\n"; fi
 
-        #Convert UTXO into mary style if UTXO is shelley/allegra style
-        if [[ ! "$(jq -r '[.[]][0].amount | type' <<< ${utxoJSON})" == "array" ]]; then utxoJSON=$(convert_UTXO "${utxoJSON}"); fi
-
 	#Calculating the total amount of lovelaces in all utxos on this address
         totalLovelaces=$(jq '[.[].amount[0]] | add' <<< ${utxoJSON})
 
         totalAssetsJSON="{}"; 	#Building a total JSON with the different assetstypes "policyIdHash.name", amount and name
         totalPolicyIDsJSON="{}"; #Holds the different PolicyIDs as values "policyIDHash", length is the amount of different policyIDs
 
-	assetsOutString="";	#This will hold the String to append on the --tx-out if assets present or it will be empty
+	assetsReturnString="";	#This will hold the String to append on the --tx-out if assets present or it will be empty
 
         #For each utxo entry, check the utxo#index and check if there are also any assets in that utxo#index
         #LEVEL 1 - different UTXOs
         for (( tmpCnt=0; tmpCnt<${txcnt}; tmpCnt++ ))
         do
-        utxoHashIndex=$(jq -r "keys_unsorted[${tmpCnt}]" <<< ${utxoJSON})
+        utxoHashIndex=$(jq -r "keys[${tmpCnt}]" <<< ${utxoJSON})
         utxoAmount=$(jq -r ".\"${utxoHashIndex}\".amount[0]" <<< ${utxoJSON})   #Lovelaces
         echo -e "Hash#Index: ${utxoHashIndex}\tAmount: ${utxoAmount}"
         assetsJSON=$(jq -r ".\"${utxoHashIndex}\".amount[1]" <<< ${utxoJSON})
@@ -143,23 +107,68 @@ echo
         txInString="${txInString} --tx-in ${utxoHashIndex}"
         done
         echo -e "\e[0m-----------------------------------------------------------------------------------------------------"
-        echo -e "Total ADA on the Address:\e[32m $(convertToADA ${totalLovelaces}) ADA / ${totalLovelaces} lovelaces \e[0m\n"
+        totalInADA=$(bc <<< "scale=6; ${totalLovelaces} / 1000000")
+        echo -e "Total ADA on the Address:\e[32m  ${totalInADA} ADA / ${totalLovelaces} lovelaces \e[0m\n"
 
-        totalPolicyIDsCnt=$(jq length <<< ${totalPolicyIDsJSON});
+	totalPolicyIDsCnt=$(jq length <<< ${totalPolicyIDsJSON});
+
+	#Calculating the amount of the asset to send out, using the given amount of using all of the assets amount
+	echo -e "\e[0mAsset-Type to send to ${toAddr}.addr:\e[33m ${assetToSend} \e[0m"
+	echo
+
+	#Currently on the source address
+	assetAmount=$(jq -r ".\"${assetToSend}\".amount" <<< ${totalAssetsJSON})
+        assetName=$(jq -r ".\"${assetToSend}\".name" <<< ${totalAssetsJSON})
+	#If there is no asset of that type, exit with an error
+	if [[ ${assetAmount} -gt 0 ]]; then
+			        echo -e "\e[0mAsset-Amount currenty on ${fromAddr}.addr:\e[32m ${assetAmount} ${assetName} \e[0m"
+        				else
+				echo -e "\n\e[35mError - This asset is not available on the source address!\e[0m\n"; exit 2;
+	fi
+
+	#If keyword ALL was used, set the amountToSend to the available amount on the source address
+	if [[ "${amountToSend^^}" == "ALL" ]]; then amountToSend=${assetAmount}; fi
+
+        echo -e "\e[0mAsset-Amount to send to ${toAddr}.addr:\e[33m ${amountToSend} ${assetName} \e[0m"
+
+        if [[ ${amountToSend} -lt 1 ]]; then  echo -e "\n\e[35mError - Please input a positive sending amount (integer)!\e[0m\n"; exit 2; fi
+
+	amountToReturn=$(( ${assetAmount} - ${amountToSend} ));	#the rest amount of the asset that stays on the source address
+
+        if [[ ${amountToReturn} -ge 0 ]]; then
+                                echo -e "\e[0mAsset-Amount that stays on the source Address:\e[32m ${amountToReturn} ${assetName} \e[0m"
+                                        else
+                                echo -e "\n\e[35mError - You can't send out more amounts of this assets than it is available on the source address!\e[0m\n"; exit 2;
+        fi
+
+	#Update the new value in the totalAssetsJSON
+	totalAssetsJSON=$( jq ". += {\"${assetToSend}\":{amount: ${amountToReturn}, name: \"${assetName}\"}}" <<< ${totalAssetsJSON})
+
+	echo
+
         totalAssetsCnt=$(jq length <<< ${totalAssetsJSON})
         if [[ ${totalAssetsCnt} -gt 0 ]]; then
-                        echo -e "\e[32m${totalAssetsCnt} Asset-Type(s) / ${totalPolicyIDsCnt} different PolicyIDs\e[0m found on the Address!\n"
+                        echo -e "\e[32m${totalAssetsCnt} Asset-Type(s) / ${totalPolicyIDsCnt} Policy-IDs \e[0m - Remaining assets on the source address\n"
                         printf "\e[0m%-70s %16s %s\n" "PolicyID.Name:" "Total-Amount:" "Name:"
-                        for (( tmpCnt=0; tmpCnt<${totalAssetsCnt}; tmpCnt++ ))
+                         for (( tmpCnt=0; tmpCnt<${totalAssetsCnt}; tmpCnt++ ))
                         do
                         assetHashName=$(jq -r "keys[${tmpCnt}]" <<< ${totalAssetsJSON})
                         assetAmount=$(jq -r ".\"${assetHashName}\".amount" <<< ${totalAssetsJSON})
                         assetName=$(jq -r ".\"${assetHashName}\".name" <<< ${totalAssetsJSON})
                         printf "\e[90m%-70s \e[32m%16s %s\e[0m\n" "${assetHashName}" "${assetAmount}" "${assetName}"
-                        if [[ ${assetAmount} -gt 0 ]]; then assetsOutString+="+${assetAmount} ${assetHashName}"; fi #only include in the sendout if more than zero
+                        if [[ ${assetAmount} -gt 0 ]]; then assetsReturnString+="+${assetAmount} ${assetHashName}"; fi #only include in the sendout if more than zero
                         done
         fi
+
+	assetsSendString="+${amountToSend} ${assetToSend}"
+
 echo
+
+#
+# Set the right rxcnt, in this case its always 2. One external destination address, one return address (sourceAddress)
+#
+
+rxcnt=2
 
 #Read ProtocolParameters
 if ${onlineMode}; then
@@ -168,28 +177,36 @@ if ${onlineMode}; then
                         protocolParametersJSON=$(jq ".protocol.parameters" <<< ${offlineJSON}); #offlinemode
                   fi
 checkError "$?"
-minOutUTXO=$(get_minOutUTXO "${protocolParametersJSON}" "${totalAssetsCnt}" "${totalPolicyIDsCnt}")
-
-#-------------------------------------
+minOutUTXO=$(get_minOutUTXO "${protocolParametersJSON}" "1" "1")  #one asset will be sent out, so total assetcount=1 and policyidcount=1
+if [[ ${amountToReturn} -gt 0 ]]; then
+ 					minReturnUTXO=$(get_minOutUTXO "${protocolParametersJSON}" "${totalAssetsCnt}" "${totalPolicyIDsCnt}") #restamount of the sending asset still on the source so full number of assets to return
+                                        else
+					minReturnUTXO=$(get_minOutUTXO "${protocolParametersJSON}" "$(( ${totalAssetsCnt} - 1 ))" "${totalPolicyIDsCnt}") #the choosen asset is transfered completly, so total number - 1
+fi
 
 
 #Generate Dummy-TxBody file for fee calculation
 txBodyFile="${tempDir}/dummy.txbody"
 rm ${txBodyFile} 2> /dev/null
-${cardanocli} ${subCommand} transaction build-raw ${nodeEraParam} ${txInString} --tx-out "${dummyShelleyAddr}+0${assetsOutString}" --invalid-hereafter ${ttl} --fee 0 --metadata-json-file ${metafile} --out-file ${txBodyFile}
+${cardanocli} ${subCommand} transaction build-raw ${nodeEraParam} ${txInString} --tx-out "${dummyShelleyAddr}+0${assetsSendString}" --tx-out "${dummyShelleyAddr}+0${assetsReturnString}" --invalid-hereafter ${ttl} --fee 0 --out-file ${txBodyFile}
 checkError "$?"
 fee=$(${cardanocli} ${subCommand} transaction calculate-min-fee --tx-body-file ${txBodyFile} --protocol-params-file <(echo ${protocolParametersJSON}) --tx-in-count ${txcnt} --tx-out-count ${rxcnt} ${magicparam} --witness-count 1 --byron-witness-count 0 | awk '{ print $1 }')
 checkError "$?"
-echo -e "\e[0mMinimum Transaction Fee for ${txcnt}x TxIn & ${rxcnt}x TxOut: \e[32m $(convertToADA ${fee}) ADA / ${fee} lovelaces \e[90m"
 
-lovelacesToSend=$(( ${totalLovelaces} - ${fee} ))
+echo -e "\e[0mMinimum Transaction Fee for ${txcnt}x TxIn & ${rxcnt}x TxOut: \e[32m ${fee} lovelaces \e[90m"
 
-echo -e "\e[0mLovelaces that will be returned to payment Address (UTXO-Sum minus fees): \e[32m $(convertToADA ${lovelacesToSend}) ADA / ${lovelacesToSend} lovelaces \e[90m (min. required ${minOutUTXO} lovelaces)"
+#
+# Set the right amount of lovelacesToSend, lovelacesToReturn and also check about sendinglimits like minUTxOValue for returning assets if available
+#
+
+echo -e "\e[0mMinimum UTXO value for sending the asset: \e[32m ${minOutUTXO} lovelaces \e[90m"
 echo
+if [[ ${lovelacesToSend} -lt ${minOutUTXO} ]]; then lovelacesToSend=${minOutUTXO}; fi
+echo -e "\e[0mLovelaces to send to ${toAddr}.addr: \e[33m ${lovelacesToSend} lovelaces \e[90m"
 
-#Checking about minimum funds in the UTX0
-if [[ ${lovelacesToSend} -lt ${minOutUTXO} ]]; then echo -e "\e[35mNot enough funds on the source Addr! Minimum UTXO value is ${minOutUTXO} lovelaces.\e[0m"; exit; fi
-
+lovelacesToReturn=$(( ${totalLovelaces} - ${fee} - ${lovelacesToSend} ))
+echo -e "\e[0mLovelaces to return to ${fromAddr}.addr: \e[32m ${lovelacesToReturn} lovelaces \e[90m"
+if [[ ${lovelacesToReturn} -lt ${minReturnUTXO} ]]; then echo -e "\e[35mError - Not enough funds on the source Addr! Minimum UTXO value to return is ${minReturnUTXO} lovelaces.\e[0m"; exit; fi
 
 txBodyFile="${tempDir}/$(basename ${fromAddr}).txbody"
 txFile="${tempDir}/$(basename ${fromAddr}).tx"
@@ -200,25 +217,19 @@ echo
 
 #Building unsigned transaction body
 rm ${txBodyFile} 2> /dev/null
-${cardanocli} ${subCommand} transaction build-raw ${nodeEraParam} ${txInString} --tx-out "${sendToAddr}+${lovelacesToSend}${assetsOutString}" --invalid-hereafter ${ttl} --fee ${fee} --metadata-json-file ${metafile} --out-file ${txBodyFile}
+${cardanocli} ${subCommand} transaction build-raw ${nodeEraParam} ${txInString} --tx-out "${sendToAddr}+${lovelacesToSend}${assetsSendString}" --tx-out "${sendFromAddr}+${lovelacesToReturn}${assetsReturnString}" --invalid-hereafter ${ttl} --fee ${fee} --out-file ${txBodyFile}
 checkError "$?"
 
 cat ${txBodyFile}
 echo
 
-echo -e "\e[0mSign the unsigned transaction body with the \e[32m${fromAddr}.skey/hwsfile\e[0m: \e[32m ${txFile} \e[90m"
+echo -e "\e[0mSign the unsigned transaction body with the \e[32m${fromAddr}.skey\e[0m: \e[32m ${txFile} \e[90m"
 echo
 
-#If payment address is a hardware wallet, use the cardano-hw-cli for the signing
-if [[ -f "${fromAddr}.hwsfile" ]]; then
-        start_HwWallet; checkError "$?"; if [ $? -ne 0 ]; then exit $?; fi
-        tmp=$(${cardanohwcli} shelley transaction sign --tx-body-file ${txBodyFile} --hw-signing-file ${fromAddr}.hwsfile ${magicparam} --out-file ${txFile} 2> /dev/stdout)
-        if [[ "${tmp^^}" == *"ERROR"* ]]; then echo -e "\e[35m${tmp}\e[0m\n"; exit 1; else echo -e "\e[32mDONE\e[0m\n"; fi
-        checkError "$?"; if [ $? -ne 0 ]; then exit $?; fi
-else
-        ${cardanocli} ${subCommand} transaction sign --tx-body-file ${txBodyFile} --signing-key-file ${fromAddr}.skey ${magicparam} --out-file ${txFile}
-        checkError "$?"; if [ $? -ne 0 ]; then exit $?; fi
-fi
+#Sign the unsigned transaction body with the SecureKey
+rm ${txFile} 2> /dev/null
+${cardanocli} ${subCommand} transaction sign --tx-body-file ${txBodyFile} --signing-key-file ${fromAddr}.skey ${magicparam} --out-file ${txFile} 
+checkError "$?"
 
 cat ${txFile}
 echo
@@ -241,7 +252,7 @@ if ask "\e[33mDoes this look good for you, continue ?" N; then
                                                                         sendToAddr: \"${sendToAddr}\",
                                                                         txJSON: ${txFileJSON} } ]" <<< ${offlineJSON})
                                 #Write the new offileFile content
-                                offlineJSON=$( jq ".history += [ { date: \"$(date -R)\", action: \"signed utxo-transaction from '${fromAddr}' to '${toAddr}' - votingMeta\" } ]" <<< ${offlineJSON})
+                                offlineJSON=$( jq ".history += [ { date: \"$(date -R)\", action: \"generated and signed utxo-token-transaction from '${fromAddr}' to '${toAddr}'\" } ]" <<< ${offlineJSON})
                                 offlineJSON=$( jq ".general += {offlineCLI: \"${versionCLI}\" }" <<< ${offlineJSON})
                                 offlineJSON=$( jq ".general += {offlineNODE: \"${versionNODE}\" }" <<< ${offlineJSON})
                                 echo "${offlineJSON}" > ${offlineFile}
@@ -255,8 +266,10 @@ if ask "\e[33mDoes this look good for you, continue ?" N; then
                                 fi
 
         fi
-
 fi
 
 
 echo -e "\e[0m\n"
+
+
+
